RecommenderBackend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pandas
pyarrow
numpy
faiss-cpu
